requirements.txt
asgiref==3.2.10
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.1
djangorestframework==3.12.4
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
openapi-codec==1.3.2
pytz==2020.1
requests==2.25.1
simplejson==3.17.2
sqlparse==0.3.1
transaction==3.0.1
uritemplate==3.0.1
urllib3==1.26.5
zope.interface==5.4.0
